backend/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.31.0
pydantic>=2.10
openai>=1.54.3
python-multipart==0.0.12
matplotlib==3.9.2
pandas==2.2.3
openpyxl==3.1.5
pillow==10.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
